fix[litemall-admin-api]: 对于上传的request body中的json数据进行校验。

+++ litemall-admin-api/src/main/java/org/linlinjava/litemall/admin/web/AdminAdminController.java
@@ -2,6 +2,7 @@ package org.linlinjava.litemall.admin.web;
 
 import org.linlinjava.litemall.admin.annotation.LoginAdmin;
 import org.linlinjava.litemall.admin.service.AdminTokenManager;
+import org.linlinjava.litemall.core.util.RegexUtil;
 import org.linlinjava.litemall.core.util.ResponseUtil;
 import org.linlinjava.litemall.core.util.bcrypt.BCryptPasswordEncoder;
 import org.linlinjava.litemall.core.validator.Order;
@@ -10,6 +11,7 @@ import org.linlinjava.litemall.db.domain.LitemallAdmin;
 import org.linlinjava.litemall.db.service.LitemallAdminService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.util.Assert;
+import org.springframework.util.StringUtils;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
 
@@ -70,25 +72,38 @@ public class AdminAdminController {
         return ResponseUtil.ok(data);
     }
 
+    private Object validate(LitemallAdmin admin) {
+        String name = admin.getUsername();
+        if(StringUtils.isEmpty(name)){
+            return ResponseUtil.badArgument();
+        }
+        if(RegexUtil.isUsername(name)){
+            return ResponseUtil.fail(402, "管理员名称不符合规定");
+        }
+        String password = admin.getPassword();
+        if(StringUtils.isEmpty(password) || password.length() < 6){
+            return ResponseUtil.fail(402, "管理员密码长度不能小于6");
+        }
+        return null;
+    }
+
     @PostMapping("/create")
     public Object create(@LoginAdmin Integer adminId, @RequestBody LitemallAdmin admin){
         if(adminId == null){
             return ResponseUtil.unlogin();
         }
+        Object error = validate(admin);
+        if(error != null){
+            return error;
+        }
 
         String username = admin.getUsername();
-        if(username == null){
-            return ResponseUtil.badArgument();
-        }
         List<LitemallAdmin> adminList = adminService.findAdmin(username);
         if(adminList.size() > 0){
             return ResponseUtil.fail(402, "管理员已经存在");
         }
 
         String rawPassword = admin.getPassword();
-        if(rawPassword == null || rawPassword.length() < 6){
-            return ResponseUtil.fail(402, "管理员密码长度不能小于6");
-        }
         BCryptPasswordEncoder encoder = new BCryptPasswordEncoder();
         String encodedPassword = encoder.encode(rawPassword);
         admin.setPassword(encodedPassword);
@@ -113,8 +128,17 @@ public class AdminAdminController {
         if(adminId == null){
             return ResponseUtil.unlogin();
         }
+        Object error = validate(admin);
+        if(error != null){
+            return error;
+        }
 
         Integer anotherAdminId = admin.getId();
+        if(anotherAdminId == null){
+            return ResponseUtil.badArgument();
+        }
+        // TODO 这里开发者需要删除以下检验代码
+        // 目前这里不允许修改超级管理员是防止演示平台上他人修改管理员密码而导致登录失败
         if(anotherAdminId == 1){
             return ResponseUtil.fail(403, "超级管理员不能修改");
         }
@@ -138,9 +162,15 @@ public class AdminAdminController {
         }
 
         Integer anotherAdminId = admin.getId();
+        if(anotherAdminId == null){
+            return ResponseUtil.badArgument();
+        }
+        // TODO 这里开发者需要删除以下检验代码
+        // 目前这里不允许删除超级管理员是防止演示平台上他人删除管理员账号而导致登录失败
         if(anotherAdminId == 1){
             return ResponseUtil.fail(403, "超级管理员不能删除");
         }
+
         adminService.deleteById(anotherAdminId);
         return ResponseUtil.ok();
     }
